server/finova/agents/first_crew_finova/requirements.txt
# Core CrewAI dependencies
crewai
crewai-tools
pydantic

# PDF handling (only for text extraction, no system dependencies needed)
PyPDF2

# LLM API clients (choose one or both)
openai>=1.0.0
anthropic>=0.25.0

# Image processing (for PDF to image conversion)
pdf2image
pillow
